🚚 Move the `Hook` enum to the `Bot` namespace
